Crate canopy_core

Core types and traits for the Canopy terminal UI library.

Re-exports

pub use error::Error;

pub use error::Result;

pub use node::EventOutcome;

pub use node::Node;

pub use state::NodeId;

pub use state::NodeName;

pub use state::NodeState;

pub use render::Render;

pub use render::RenderBackend;

pub use state::StatefulNode;

pub use state::StatefulNode as StatefulNodeTrait;

pub use geom;

pub use commands::*;

Modules

backend

commands

cursor

dump

error

event

node

path

render

script

state

style

tree

Utilities for working with a Canopy node tree.

tutils

Macros

buf

A helper macro to create buffers for the termbuf match assertions.

rgb

Macro to create a Color from a hex string at compile time

Structs

Binder

Binder provides an ergonomic way to specify a set of key bindings using a builder patttern.

Canopy

Expanse

An Expanse is a rectangle that has a width and height but no location. This is useful when we want to deal with Rects abstractly, or when we want to madate that the location of a Rect is (0, 0).

FocusableNode

Information about a focusable node

InputMap

The InputMap struct manages the global set of key and mouse bindings for the app.

InputMode

A InputMode contains a set of bound keys and mouse actions.

Layout

The Layout struct provides operations that a node can perform on children during its layout phase.

Point

Poller

The Poller is responsible for scheduling poll events for nodes.

Rect

A rectangle

TermBuf

ViewPort

A ViewPort manages the size of a node and its projection onto the screen. In many ways, this is the core of Canopy, and the viewport structure and its constraints determines many aspects of Canopy’s layout and rendering behavior.

ViewStack

A stack of viewports that manages nested view transformations.

Enums

Direction

Input

Traits

Context

The API exposed to nodes by Canopy.

DefaultBindings

Loader

A trait that allows widgets to perform recursive initialization of themselves and their children. The most common use for this trait is to load the command sets from a node tree.

Functions

collect_focusable_nodes

Collect all focusable nodes with their screen rectangles

find_focus_target

Find the best focus target in the specified direction

find_focused_node

Find the currently focused node and its screen rectangle

Attribute Macros

command

Mark a method as a command. This macro should be used to decorate methods in an impl block that uses the derive_commands macro. A number of optional arguments can be passed:

derive_commands

Derive an implementation of the CommandNode trait. This macro should be added to the impl block of a struct. All methods that are annotated with command are added as commands, with their doc comments as the command documentation.

Derive Macros

DeriveStatefulNode

Derive an implementation of the StatefulNode trait for a struct. The struct should have a self.state attribute of type NodeState.

StatefulNode

Derive an implementation of the StatefulNode trait for a struct. The struct should have a self.state attribute of type NodeState.