canopy_core

Module state

Structs

NodeId

A unique ID for a node.

NodeName

A node name, which consists of lowercase ASCII alphanumeric characters, plus underscores.

NodeState

An opaque structure that Canopy uses to track node state. Each Node has to keep a NodeState structure, and offer it up through the Node::state() method on request.

Traits

StatefulNode

The interface implemented by all nodes that track state.

Functions

valid_nodename

valid_nodename_char