canopy_core::state

Trait StatefulNode

pub trait StatefulNode {

    // Required methods
    fn name(&self) -> NodeName;
    fn state(&self) -> &NodeState;
    fn state_mut(&mut self) -> &mut NodeState;

    // Provided methods
    fn hide(&mut self) { ... }
    fn unhide(&mut self) { ... }
    fn is_hidden(&self) -> bool { ... }
    fn vp(&self) -> ViewPort { ... }
    fn id(&self) -> NodeId { ... }
    fn is_initialized(&self) -> bool { ... }
    fn set_canvas(&mut self, sz: Expanse) { ... }
    fn set_view(&mut self, view: Rect) { ... }
    fn fill(&mut self, sz: Expanse) -> Result<()> { ... }
    fn wrap(&mut self, child: ViewPort) -> Result<()> { ... }
    fn fit_size(&mut self, size: Expanse, view_size: Expanse) { ... }
    fn scroll_to(&mut self, x: u32, y: u32) { ... }
    fn scroll_by(&mut self, x: i32, y: i32) { ... }
    fn page_up(&mut self) { ... }
    fn page_down(&mut self) { ... }
    fn scroll_up(&mut self) { ... }
    fn scroll_down(&mut self) { ... }
    fn scroll_left(&mut self) { ... }
    fn scroll_right(&mut self) { ... }

}

The interface implemented by all nodes that track state.

Required Methods

fn name(&self) -> NodeName

The name of this node, used for debugging and command dispatch.

fn state(&self) -> &NodeState

Get a reference to the node’s state object.

fn state_mut(&mut self) -> &mut NodeState

Get a mutable reference to the node’s state object.

Provided Methods

fn hide(&mut self)

Hides the element and all its descendants from rendering. The nodes are still included in the tree.

fn unhide(&mut self)

Hides the element

fn is_hidden(&self) -> bool

Is this element hidden?

fn vp(&self) -> ViewPort

Get the node’s ViewPort.

fn id(&self) -> NodeId

A unique ID for this node.

fn is_initialized(&self) -> bool

Has this node been initialized? That is, has its poll function been called for the first time to schedule future polls.

fn set_canvas(&mut self, sz: Expanse)

Set our canvas size.

fn set_view(&mut self, view: Rect)

Set our view position and size.

fn fill(&mut self, sz: Expanse) -> Result<()>

Set both the canvas size and the view to fill the target size.

fn wrap(&mut self, child: ViewPort) -> Result<()>

Wrap around a child by seting both our canvas size and view to equal to its view rectangle.

fn fit_size(&mut self, size: Expanse, view_size: Expanse)

fn scroll_to(&mut self, x: u32, y: u32)

Scroll the view to the specified position.

fn scroll_by(&mut self, x: i32, y: i32)

Scroll the view by the given offsets.

fn page_up(&mut self)

Scroll the view up by the height of the view rectangle.

fn page_down(&mut self)

Scroll the view down by the height of the view rectangle.

fn scroll_up(&mut self)

Scroll the view up by one line.

fn scroll_down(&mut self)

Scroll the view down by one line.

fn scroll_left(&mut self)

Scroll the view left by one line.

fn scroll_right(&mut self)

Scroll the view right by one line.