canopy_core::tree

Function preorder

pub fn preorder<T>(
    e: &mut dyn Node,
    f: &mut dyn FnMut(&mut dyn Node) -> Result<Walk<T>>,
) -> Result<Walk<T>>

Walk::Skip prunes all children of the current node from the traversal.

Walk::Handle stops the traversal and the contained value is returned.

Any error return stops the traversal and the error is returned.