canopy_core

Struct TermBuf

pub struct TermBuf { /* private fields */ }

A 2D terminal buffer of styled cells.

Implementations

impl TermBuf

pub fn new(size: impl Into<Expanse>, ch: char, style: Style) -> Self

Construct a buffer filled with the given character and style.

pub fn empty_with_style(size: impl Into<Expanse>, style: Style) -> Self

Create an empty TermBuf filled with NULL characters.

pub fn empty(size: impl Into<Expanse>) -> Self

Create an empty TermBuf filled with NULL characters.

pub fn copy(&mut self, src: &Self, rect: Rect)

Copy non-NULL characters from a rectangle of another TermBuf into this one

pub fn copy_to_rect(&mut self, src: &Self, dest_rect: Rect)

Copy non-NULL characters from a source TermBuf into a destination rectangle

pub fn size(&self) -> Expanse

Return the buffer size.

pub fn rect(&self) -> Rect

Return the buffer bounds as a rectangle.

pub fn fill(&mut self, style: &Style, r: Rect, ch: char)

Fill a rectangle with a glyph and style.

pub fn fill_empty(&mut self, ch: char, style: &Style)

Fill all empty (NULL) cells with the given character and style.

pub fn solid_frame(&mut self, style: &Style, f: Frame, ch: char)

Fill the frame outline with a glyph and style.

pub fn text(&mut self, style: &Style, l: Line, txt: &str)

Draw text clipped to the given line.

pub fn get(&self, p: Point) -> Option<&Cell>

Get a cell by position.

pub fn diff<R: RenderBackend>(&self, prev: &Self, backend: &mut R) -> Result<()>

Diff this terminal buffer against a previous state, emitting changes to the provided render backend.

pub fn render<R: RenderBackend>(&self, backend: &mut R) -> Result<()>

Render this terminal buffer in full using the provided backend, batching runs of text with the same style.